requirements.txt
# --- Core Azure AI SDKs ---
azure-ai-documentintelligence==1.0.2
openai==1.74.0
azure-identity==1.21.0
python-dotenv==1.1.0
gradio==5.25.0
fastapi==0.115.12
uvicorn==0.34.1
langchain==0.3.23            # Framework to help with RAG pipelines (chunking, prompts, chains)
langchain-openai==0.3.12     # LangChain integrations for OpenAI
langchain-community==0.3.21 # LangChain community integrations (might be needed)
beautifulsoup4==4.13.3        # For parsing HTML from phase2_data
faiss-cpu==1.10.0           # For efficient vector similarity search (CPU version)
tiktoken==0.9.0            # For token counting with OpenAI models
numpy
pydantic
langdetect